Rearrange code, add instances.
Moved Control.Eff and most of Control.Eff.Lifted into
Control.Eff.Internal and added MonadBase and MonadBaseControl instances.
Not tested thoroughly yet.
