+++ extensible-effects.cabal
@@ -88,7 +88,8 @@ library
                        Data.OpenUnion
 
   -- Modules included in this library but not exported.
-  other-modules:       Data.FTCQueue
+  other-modules:       Control.Eff.Internal
+                       Data.FTCQueue
   if flag(force-openunion-51)
     cpp-options:       -DFORCE_OU51
 
@@ -134,10 +135,10 @@ library
 
   -- Other library packages from which modules are imported.
   build-depends:       base >= 4.7 && < 4.11
-                       -- For MonadIO instance
-                       , transformers >= 0.3 && < 0.6
-                       -- For MonadBase instance
-                       , transformers-base == 0.4.*
+                       -- For MonadBase
+               ,       transformers-base == 0.4.*
+                       -- For MonadBaseControl
+               ,       monad-control >= 1.0 && < 1.1
 
   -- Directories containing source files.
   hs-source-dirs:      src
